Added can.js and loading animation
FIxed a number of issues pertaining to the html5 boiler plate template.
The content now only shows when all the javascript is loaded.

Integration with can.js has now begun, and I will soon be adding full
template support for AJAX loading, and a back-end application for
quickly adding content.  This will NOT be a substitute for a CMS, and
rather, will be a way for people to quickly add content to their live
site.